requirements.txt
certifi==2018.4.16
chardet==3.0.4
coreapi==2.3.3
coreschema==0.0.4
Django==1.11.15
django-extensions==2.1.0
django-filter==1.1
django-rest-swagger==0.3.10
djangorestframework==3.3.3
idna==2.7
itypes==1.1.0
Jinja2==2.10
Markdown==2.5
MarkupSafe==1.0
openapi-codec==1.3.2
pytz==2018.5
PyYAML==3.13
requests==2.19.1
simplejson==3.16.0
six==1.11.0
typing==3.6.4
uritemplate==3.0.0
urllib3==1.23
Werkzeug==0.14.1
